+++ ApplicationTier.Infrastructure/Data/ViewModel/ShoppingCartVM.cs
@@ -5,7 +5,7 @@ namespace RocketEshop.Infrastructure.Data.ViewModel
     public class ShoppingCartVM
     {
         public ShoppingCart ShoppingCart { get; set; }
-        public double ShoppingCartTotal { get; set; }
+        public decimal ShoppingCartTotal { get; set; }
 
     }
 }
